修改脚本

+++ webhook/scripts/back.sh
@@ -1,9 +1,19 @@
 pkill maile-backend
-rm /root/back/maile-backend
+
+if [ -f /root/back/maile-backend ]; then
+    rm /root/back/maile-backend
+fi
+
 go get -u github.com/mlgaku/back
 cd $GOPATH/src/github.com/mlgaku/back
 go build main.go
-mkdir /root/back
+
+if [ ! -d /root/back ]; then
+    mkdir /root/back
+fi
+
 mv main /root/back/maile-backend
 cd /root/back
 ./maile-backend > /dev/null 2>&1 &
+
+echo 'Ok'
